requirements.txt
# Backend Dependencies for Todo Full-Stack Application

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
sqlmodel>=0.0.14
alembic>=1.13.1

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Authentication
bcrypt>=4.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpx>=0.26.0

# Development
python-dotenv>=1.0.0
black>=23.12.1
isort>=5.13.2
mypy>=1.8.0
